requirements.txt
acme==0.31.0
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bleach==3.1.0
captcha==0.3
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.6.1
distro==1.3.0
distro-info==0.21
entrypoints==0.3
Faker==2.0.2
Flask==1.1.1
flask==1.1.1
Flask-Caching==1.7.2
Flask-Limiter==1.0.1
Flask-Session==0.3.1
flask-session-captcha==1.1.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==2.4.1
flask-sqlalchemy==2.4.1
future==0.16.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.1
josepy==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
limits==1.3
Markdown==3.1.1
MarkupSafe==1.1.1
mock==2.0.0
parsedatetime==2.4
pbr==4.2.0
Pillow==6.1.0
psycopg2==2.7.7
pycrypto==2.6.1
PyGObject==3.30.4
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.1
python-apt==1.8.4
python-dateutil==2.8.0
pytz==2019.1
pyxdg==0.25
requests==2.21.0
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.8
text-unidecode==1.3
unattended-upgrades==0.1
urllib3==1.24.1
uWSGI==2.0.18
webencodings==0.5.1
Werkzeug==0.16.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
